requirements.txt
Flask==2.3.2
langchain==0.0.247
playsound==1.3.0
python-dotenv==1.0.0
Requests==2.31.0
